Added missing RV64 instructions, memory fixes

+++ src/cpu/memory/raw_table_memory.rs
@@ -8,8 +8,7 @@ use super::{
 
 #[derive(Clone)]
 pub struct RawTableMemoryPageStorage {
-    pages: [Option<Box<Page>>;
-        (((MEMORY_SIZE as u64) + (PAGE_SIZE as u64) - 1) / (PAGE_SIZE as u64)) as usize],
+    pages: [Option<Box<Page>>; (((MEMORY_SIZE) + (PAGE_SIZE) - 1) / (PAGE_SIZE)) as usize],
 }
 
 impl Debug for RawTableMemoryPageStorage {
